refactor abs,cbrt,ceil, clip
