+++ daft/daft/__init__.pyi
@@ -1411,16 +1411,13 @@ class PySeries:
     def agg_list(self) -> PySeries: ...
     def agg_set(self) -> PySeries: ...
     def cast(self, dtype: PyDataType) -> PySeries: ...
-    def ceil(self) -> PySeries: ...
     def floor(self) -> PySeries: ...
     def sign(self) -> PySeries: ...
     def signum(self) -> PySeries: ...
     def negate(self) -> PySeries: ...
     def negative(self) -> PySeries: ...
     def round(self, decimal: int) -> PySeries: ...
-    def clip(self, min: PySeries, max: PySeries) -> PySeries: ...
     def sqrt(self) -> PySeries: ...
-    def cbrt(self) -> PySeries: ...
     def sin(self) -> PySeries: ...
     def cos(self) -> PySeries: ...
     def tan(self) -> PySeries: ...
