+++ src/daft-core/src/python/series.rs
@@ -141,10 +141,6 @@ impl PySeries {
         Ok(self.series.floor_div(&other.series)?.into())
     }
 
-    pub fn ceil(&self) -> PyResult<Self> {
-        Ok(self.series.ceil()?.into())
-    }
-
     pub fn floor(&self) -> PyResult<Self> {
         Ok(self.series.floor()?.into())
     }
@@ -165,16 +161,9 @@ impl PySeries {
         }
         Ok(self.series.round(decimal)?.into())
     }
-    pub fn clip(&self, min: &Self, max: &Self) -> PyResult<Self> {
-        Ok(self.series.clip(&min.series, &max.series)?.into())
-    }
 
     pub fn sqrt(&self) -> PyResult<Self> {
         Ok(self.series.sqrt()?.into())
-    }
-
-    pub fn cbrt(&self) -> PyResult<Self> {
-        Ok(self.series.cbrt()?.into())
     }
 
     pub fn sin(&self) -> PyResult<Self> {
