+++ src/daft-core/src/series/ops/mod.rs
@@ -10,9 +10,6 @@ pub mod between;
 pub mod binary;
 pub mod broadcast;
 pub mod cast;
-pub mod cbrt;
-pub mod ceil;
-pub mod clip;
 pub mod comparison;
 pub mod concat;
 pub mod downcast;
