+++ src/daft-functions/src/numeric/cbrt.rs
@@ -26,7 +26,17 @@ impl ScalarUDF for Cbrt {
     }
 
     fn evaluate(&self, inputs: &[Series]) -> DaftResult<Series> {
-        evaluate_single_numeric(inputs, Series::cbrt)
+        evaluate_single_numeric(inputs, |s| {
+            let casted_dtype = s.to_floating_data_type()?;
+            let casted_self = s
+                .cast(&casted_dtype)
+                .expect("Casting numeric types to their floating point analogues should not fail");
+            match casted_dtype {
+                DataType::Float32 => Ok(casted_self.f32().unwrap().cbrt()?.into_series()),
+                DataType::Float64 => Ok(casted_self.f64().unwrap().cbrt()?.into_series()),
+                _ => unreachable!(),
+            }
+        })
     }
 }
